utils/metadata_update/min_requirements.txt
attrs==23.2.0
certifi==2023.11.17
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
et-xmlfile==1.1.0
fiona==1.9.5
geopandas==0.13.2
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==1.5.3
pip==23.3.1
psycopg2==2.9.9
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.4
setuptools==68.2.2
shapely==2.0.1
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.9.0
wheel==0.41.2
